refactor(archives): fold beat generators into buildWarNarrative; clamp day labels to lastDay
