+++ src/features/archives/buildWarNarrative.mjs
@@ -8,19 +8,178 @@ import {
     pickVariant,
     PHRASE_KEY,
 } from '@/features/archives/narrativePhrasing.mjs';
-import { buildPlayerBeats } from '@/features/archives/playerBeats.mjs';
-import { buildConquestBeats } from '@/features/archives/conquestBeats.mjs';
-import { buildNumbersBeat } from '@/features/archives/numbersBeat.mjs';
-
-const SECONDS_PER_DAY = 86400;
+import { formatNumber } from '@/shared/utils/format/formatNumber.mjs';
+import { dayOf, SECONDS_PER_DAY } from '@/shared/utils/game/warClock.mjs';
 
 function factionName(enemy) {
     return (factions[enemy]?.name ?? 'Unknown forces').replace(/^The\s+/i, '');
 }
 
-function dayOf(time, warStart) {
-    const day = Math.floor((time - warStart) / SECONDS_PER_DAY) + 1;
-    return day < 1 ? 1 : day;
+const GATES_THRESHOLD = 0.9; // "at the gates" — homeworld-assault range
+const SURGE_FACTOR = 1.4; // a peak ≥ 1.4× baseline is a "rally"
+const COLLAPSE_FACTOR = 0.6; // a trough ≤ 0.6× baseline is "the front grows quiet"
+
+function median(nums) {
+    if (nums.length === 0) return 0;
+    const s = [...nums].sort((a, b) => a - b);
+    const mid = Math.floor(s.length / 2);
+    return s.length % 2 ? s[mid] : (s[mid - 1] + s[mid]) / 2;
+}
+
+/**
+ * Player surge/collapse beats from the per-bucket player timeseries. The single
+ * most extreme surge and the single most extreme collapse (past the opening
+ * ramp) are emitted when they clear their thresholds — at most 2 beats.
+ *
+ * @param {Array<{ time:number, day:number, total:number }>} playerTimeseries - per-bucket player count samples
+ * @param {number} season - phrasing seed
+ * @returns {Array<{ time:number, day:number, kind:'surge'|'collapse', text:string }>}
+ */
+function buildPlayerBeats(playerTimeseries, season) {
+    const series = playerTimeseries ?? [];
+    if (series.length < 2) return [];
+
+    const baseline = median(series.map((p) => p.total));
+    if (baseline <= 0) return [];
+
+    /** @type {Array<{ time:number, day:number, kind:'surge'|'collapse', text:string }>} */
+    const beats = [];
+
+    // Surge: global max, anywhere.
+    const peak = series.reduce((a, b) => (b.total > a.total ? b : a));
+    if (peak.total >= SURGE_FACTOR * baseline) {
+        beats.push({
+            time: peak.time,
+            day: peak.day,
+            kind: 'surge',
+            text: pickVariant(
+                PHRASES.surge,
+                season,
+                peak.time | 0,
+            )(formatNumber(peak.total)),
+        });
+    }
+
+    // Collapse: global min, skipping the first (opening-ramp) bucket.
+    const tail = series.slice(1);
+    const trough = tail.reduce((a, b) => (b.total < a.total ? b : a));
+    if (trough.total <= COLLAPSE_FACTOR * baseline) {
+        beats.push({
+            time: trough.time,
+            day: trough.day,
+            kind: 'collapse',
+            text: pickVariant(
+                PHRASES.collapse,
+                season,
+                trough.time | 0,
+            )(formatNumber(trough.total)),
+        });
+    }
+
+    return beats;
+}
+
+/**
+ * Offensive conquest milestones from the campaign snapshots. `points/points_max`
+ * is Super Earth's conquest progress toward the enemy homeworld (high = SE
+ * winning — verified vs computeMapState + the HD1 API). Emits at most two beats:
+ *   - breakthrough: first time any faction's frac first crosses GATES_THRESHOLD
+ *   - first homeworld falls: first time any faction first reads 'defeated'
+ *
+ * @param {Array<{ time:number, data:Array<{ points:number, status:string }|null> }>} snapshots - data[enemy] positional (index = faction id). Through getCampaign every entry is [f0,f1,f2] non-null (partial buckets are filtered); the null guard below is defensive only.
+ * @param {{ points:number[] }} pointsMax - Maximum points array indexed by faction
+ * @param {number} season - Season number for narrative phrase selection
+ * @param {number} warStart - Unix-seconds anchor for day 1 (war start).
+ * @returns {Array<{ time:number, day:number, kind:'conquest', text:string }>} Conquest milestone beats
+ */
+function buildConquestBeats(snapshots, pointsMax, season, warStart) {
+    const snaps = snapshots ?? [];
+    const maxes = pointsMax?.points ?? [];
+    if (snaps.length === 0) return [];
+
+    let breakthrough = null; // first snapshot any faction crosses the gates
+    let firstFall = null; // first snapshot any faction is defeated
+
+    for (const snap of snaps) {
+        const data = snap.data ?? [];
+        for (let enemy = 0; enemy < data.length; enemy++) {
+            const s = data[enemy];
+            if (!s) continue; // factions[i] may be null pre-introduction
+            const max = maxes[enemy] || 0;
+            const frac =
+                s.status === 'defeated' ? 1
+                : max > 0 ? s.points / max
+                : 0;
+            if (!breakthrough && frac >= GATES_THRESHOLD) {
+                breakthrough = { time: snap.time, enemy };
+            }
+            if (!firstFall && s.status === 'defeated') {
+                firstFall = { time: snap.time, enemy };
+            }
+        }
+        if (breakthrough && firstFall) break;
+    }
+
+    /** @type {Array<{ time:number, day:number, kind:'conquest', text:string }>} */
+    const beats = [];
+    if (breakthrough) {
+        beats.push({
+            time: breakthrough.time,
+            day: dayOf(breakthrough.time, warStart),
+            kind: 'conquest',
+            text: pickVariant(
+                PHRASES.breakthrough,
+                season,
+                breakthrough.enemy,
+            )(factionName(breakthrough.enemy)),
+        });
+    }
+    if (firstFall) {
+        beats.push({
+            time: firstFall.time,
+            day: dayOf(firstFall.time, warStart),
+            kind: 'conquest',
+            text: pickVariant(
+                PHRASES.homeworldFalls,
+                season,
+                firstFall.enemy + 10,
+            )(factionName(firstFall.enemy)),
+        });
+    }
+
+    // Dedupe: same faction reaching the gates and falling on the same day reads
+    // as one moment — keep the "falls" beat.
+    if (
+        beats.length === 2 &&
+        breakthrough &&
+        firstFall &&
+        breakthrough.enemy === firstFall.enemy &&
+        beats[0].day === beats[1].day
+    ) {
+        return [beats[1]];
+    }
+    return beats;
+}
+
+/**
+ * One "war by the numbers" beat from the season telemetry totals, anchored at
+ * `lastTime` (the last event) so the orchestrator can order it just before the
+ * closing outcome beat. Returns `null` for telemetry-less seasons.
+ *
+ * @param {{ kills:number, missions:number, accidentals:number } | null} telemetry season telemetry totals
+ * @param {number} lastTime last event timestamp for anchoring
+ * @param {number} day season day number
+ * @param {number} season season number for phrase variant selection
+ * @returns {{ time:number, day:number, kind:'numbers', text:string } | null} beat object or null if no telemetry
+ */
+function buildNumbersBeat(telemetry, lastTime, day, season) {
+    if (!telemetry) return null;
+    const text = pickVariant(PHRASES.numbers, season, PHRASE_KEY.numbers)(
+        formatNumber(telemetry.kills),
+        formatNumber(telemetry.missions),
+        formatNumber(telemetry.accidentals),
+    );
+    return { time: lastTime, day, kind: 'numbers', text };
 }
 
 /** One Ministry-voice field report for a resolved event, via seeded phrasing. */
@@ -192,7 +351,12 @@ export function buildWarNarrative(data, telemetry = null) {
     // NEW highlight beats (features 2-4) — each carries a `kind`; clamp time to
     // lastTime so none sorts after the closing outcome beat.
     for (const pb of buildPlayerBeats(data?.playerTimeseries ?? [], season)) {
-        beats.push({ ...pb, time: Math.min(pb.time, lastTime), order: seq++ });
+        beats.push({
+            ...pb,
+            time: Math.min(pb.time, lastTime),
+            day: Math.min(pb.day, lastDay),
+            order: seq++,
+        });
     }
     for (const cb of buildConquestBeats(
         data?.snapshots ?? [],
@@ -200,7 +364,12 @@ export function buildWarNarrative(data, telemetry = null) {
         season,
         warStart,
     )) {
-        beats.push({ ...cb, time: Math.min(cb.time, lastTime), order: seq++ });
+        beats.push({
+            ...cb,
+            time: Math.min(cb.time, lastTime),
+            day: Math.min(cb.day, lastDay),
+            order: seq++,
+        });
     }
 
     // Outcome (caps the chronicle) + the numbers beat just before it.
